+++ src/wxMathML.lisp
@@ -104,9 +104,11 @@
 ;; Tell maxima how to send a list of manual topics to show to us
 (defun display-frontend-topics (topiclist)
   (format t "<html-manual-keywords>")
-  (mapcar #'(lambda (&rest x) (format t "<keyword>~a</keyword>" (first (second (first x))))) topiclist)
-  (format t "</html-manual-keywords>~%")
-  )
+  (mapcar #'(lambda (&rest x) (format t "<keyword>~a</keyword>"
+				      (wxxml-fix-string
+				       (first (second (first x))))))
+	  topiclist)
+  (format t "</html-manual-keywords>~%"))
 
 ;; Escapes all chars that need escaping in XML
 ;;